modules/asr/requirements.txt
# ASR 模块依赖
# SenseVoice 语音识别

# FastAPI 框架
fastapi>=0.111.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PyTorch
torch>=2.0.0
torchaudio>=2.0.0

# SenseVoice 相关
funasr>=1.1.3
modelscope>=1.9.0
huggingface_hub>=0.19.0

# 音频处理
soundfile>=0.12.0
librosa>=0.10.0

# 配置管理
pyyaml>=6.0

# 工具库
numpy>=1.24.0,<2.0.0

